pea-sim-parametric/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pea-sim-parametric"
version = "0.1.0"
description = "Parametric Monte Carlo simulator for PEA portfolios"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "OpenAI Codex"}]
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "typer>=0.9",
  "matplotlib>=3.7",
]

[project.optional-dependencies]
test = ["pytest>=7.4"]

[project.scripts]
pea-sim = "pea_sim.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
